+++ resources/js/tool.js
@@ -1,6 +1,4 @@
 Nova.booting((Vue, router) => {
-    Vue.config.devtools = true;
-
     router.addRoutes([
         {
             name: 'nova-filemanager',