+++ src/Http/Controllers/FilemanagerToolController.php
@@ -5,6 +5,7 @@ namespace Infinety\Filemanager\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controller;
 use Infinety\Filemanager\Http\Services\FileManagerService;
+use Laravel\Nova\Http\Requests\NovaRequest;
 
 class FilemanagerToolController extends Controller
 {
@@ -32,6 +33,16 @@ class FilemanagerToolController extends Controller
     /**
      * @param Request $request
      */
+    public function getDataField($resource, $attribute, NovaRequest $request)
+    {
+        $filter = $this->getFilemanagerFieldFilter($attribute, $request);
+
+        return $this->service->ajaxGetFilesAndFolders($request, $filter);
+    }
+
+    /**
+     * @param Request $request
+     */
     public function createFolder(Request $request)
     {
         return $this->service->createFolderOnPath($request->folder, $request->current);
@@ -50,7 +61,7 @@ class FilemanagerToolController extends Controller
      */
     public function upload(Request $request)
     {
-        return $this->service->uploadFile($request->file, $request->current);
+        return $this->service->uploadFile($request->file, $request->current, $request->visibility);
     }
 
     /**
@@ -67,5 +78,22 @@ class FilemanagerToolController extends Controller
     public function removeFile(Request $request)
     {
         return $this->service->removeFile($request->file);
+    }
+
+    /**
+     * @param NovaRequest $request
+     */
+    private function getFilemanagerFieldFilter($attribute, NovaRequest $request)
+    {
+        $fields = $request->newResource()->fields($request);
+        foreach ($fields as $field) {
+            if ($field->attribute == $attribute) {
+                if (isset($field->meta['filterBy'])) {
+                    return $field->meta['filterBy'];
+                }
+            }
+        }
+
+        return false;
     }
 }