+++ src/Http/Exceptions/InvalidConfig.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Infinety\Filemanager\Exceptions;
+
+use Exception;
+
+class InvalidConfig extends Exception
+{
+    public static function driverNotSupported()
+    {
+        return new static('Driver not supported. Please check your configuration');
+    }
+}