+++ src/Http/Services/GetFiles.php
@@ -3,7 +3,6 @@
 namespace Infinety\Filemanager\Http\Services;
 
 use GuzzleHttp\Client;
-use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\Storage;
 
 trait GetFiles
@@ -31,14 +30,18 @@ trait GetFiles
 
         $files = [];
 
-        $minutes = env('FILEMANAGER_CACHE', 5);
+        $cacheTime = config('filemanager.cache', false);
 
         foreach ($filesData as $file) {
             $id = $this->generateId($file);
 
-            $fileData = Cache::remember($id, $minutes, function () use ($file, $id) {
-                return $this->getFileData($file, $id);
-            });
+            if ($cacheTime) {
+                $fileData = cache()->remember($id, $cacheTime, function () use ($file, $id) {
+                    return $this->getFileData($file, $id);
+                });
+            } else {
+                $fileData = $this->getFileData($file, $id);
+            }
 
             $files[] = $fileData;
         }
@@ -48,7 +51,7 @@ trait GetFiles
             $files = $this->filterData($files, $filter);
         }
 
-        return $this->orderData($files, $order);
+        return $this->orderData($files, $order, config('filemanager.direction', 'asc'));
     }
 
     /**
@@ -57,13 +60,14 @@ trait GetFiles
      */
     public function getFileData($file, $id)
     {
-        if (! $this->isDot($file) && ! $this->exceptExtensions->contains($file['extension']) && ! $this->exceptFolders->contains($file['basename']) && ! $this->exceptFiles->contains($file['basename']) && $this->accept($file)) {
+        if (!$this->isDot($file) && !$this->exceptExtensions->contains($file['extension']) && !$this->exceptFolders->contains($file['basename']) && !$this->exceptFiles->contains($file['basename']) && $this->accept($file)) {
             $fileInfo = [
                 'id'         => $id,
                 'name'       => trim($file['basename']),
                 'path'       => $this->cleanSlashes($file['path']),
                 'type'       => $file['type'],
                 'mime'       => $this->getFileType($file),
+                'ext'        => (isset($file['extension'])) ? $file['extension'] : false,
                 'size'       => ($file['size'] != 0) ? $file['size'] : 0,
                 'size_human' => ($file['size'] != 0) ? $this->formatBytes($file['size'], 0) : 0,
                 'thumb'      => $this->getThumbFile($file),
@@ -78,7 +82,7 @@ trait GetFiles
 
             if ($fileInfo['mime'] == 'image') {
                 list($width, $height) = $this->getImageDimesions($file);
-                if (! $width == false) {
+                if (!$width == false) {
                     $fileInfo['dimensions'] = $width.'x'.$height;
                 }
             }
@@ -99,33 +103,21 @@ trait GetFiles
     {
         $folders = $files->where('type', 'dir');
         $items = $files->where('type', 'file');
-        $filtered = $items->filter(function ($item) use ($filter) {
-            switch ($filter) {
-                case 'image':
-                    if (str_contains($item->mime, 'image')) {
+
+        $filters = config('filemanager.filters', []);
+        if (count($filters) > 0) {
+            $filters = array_change_key_case($filters);
+
+            if (isset($filters[$filter])) {
+                $filteredExtensions = $filters[$filter];
+
+                $filtered = $items->filter(function ($item) use ($filteredExtensions) {
+                    if (in_array($item->ext, $filteredExtensions)) {
                         return $item;
                     }
-                    break;
-                case 'audio':
-                    if (str_contains($item->mime, 'audio')) {
-                        return $item;
-                    }
-                    break;
-                case 'video':
-                    if (str_contains($item->mime, 'video')) {
-                        return $item;
-                    }
-                    break;
-                case 'documents':
-                    if (str_contains($item->mime, 'word') || str_contains($item->mime, 'excel') || str_contains($item->mime, 'pdf') || str_contains($item->mime, 'plain') || str_contains($item->mime, 'rtf') || str_contains($item->mime, 'text')) {
-                        return $item;
-                    }
-                    break;
-                case 'all':
-                    return $item;
-                    break;
+                });
             }
-        });
+        }
 
         return $folders->merge($filtered);
     }
@@ -214,7 +206,7 @@ trait GetFiles
      */
     public function getAppend()
     {
-        if (in_array(env('FILEMANAGER_DISK_DRIVER', $this->disk), $this->cloudDisks)) {
+        if (in_array(config('filemanager.disk'), $this->cloudDisks)) {
             return '';
         }
 
@@ -229,12 +221,13 @@ trait GetFiles
     public function getFileType($file)
     {
         $mime = $this->storage->getMimetype($file['path']);
+        $extension = $file['extension'];
 
         if (str_contains($mime, 'directory')) {
             return 'dir';
         }
 
-        if (str_contains($mime, 'image')) {
+        if (str_contains($mime, 'image') || $extension == 'svg') {
             return 'image';
         }
 
@@ -303,18 +296,14 @@ trait GetFiles
     public function getThumb($file, $folder = false)
     {
         $mime = $this->storage->getMimetype($file['path']);
+        $extension = $file['extension'];
 
         if (str_contains($mime, 'directory')) {
             return false;
         }
 
-        if (str_contains($mime, 'image')) {
-            // return $this->storage->url($file['basename']);
-            // if ($folder) {
-            //     return '/'.$folder.DIRECTORY_SEPARATOR.$file['basename'];
-            // }
-
-            if (in_array(env('FILEMANAGER_DISK_DRIVER', $this->disk), $this->cloudDisks)) {
+        if (str_contains($mime, 'image') || $extension == 'svg') {
+            if (method_exists($this->storage, 'put')) {
                 return $this->storage->url($file['path']);
             }
 
@@ -337,10 +326,8 @@ trait GetFiles
             return getimagesize($this->storage->path($file['path']));
         }
 
-        if (in_array(env('FILEMANAGER_DISK_DRIVER', $this->disk), $this->cloudDisks)) {
+        if (in_array(config('filemanager.disk'), $this->cloudDisks)) {
             return false;
-
-            return $this->getImageDimesionsFromCloud($file);
         }
 
         return false;
@@ -384,10 +371,10 @@ trait GetFiles
     public function normalizeFiles($files)
     {
         foreach ($files as $key => $file) {
-            if (! isset($file['extension'])) {
+            if (!isset($file['extension'])) {
                 $files[$key]['extension'] = null;
             }
-            if (! isset($file['size'])) {
+            if (!isset($file['size'])) {
                 $files[$key]['size'] = null;
             }
         }